BBs saved to map

+++ Tron3k/RenderPipeline/ContentManager.cpp
@@ -110,25 +110,6 @@ void ContentManager::renderChunks(GLuint shader, GLuint shaderLocation, GLuint t
 	glActiveTexture(GL_TEXTURE0 +1 );
 	glBindTexture(GL_TEXTURE_2D, textures[0].textureID);
 	testMap.render(shader, shaderLocation);
-	//for (size_t i = 0; i < meshes.size(); i++)
-	//{
-	//	glProgramUniformMatrix4fv(shader, shaderLocation, 1, GL_FALSE, (GLfloat*)meshes[i].getWorld());
-	//
-	//	glProgramUniform1i(shader, textureLocation, 0);
-	//	glProgramUniform1i(shader, normalLocation, 1);
-	//	glActiveTexture(GL_TEXTURE0);
-	//	glBindTexture(GL_TEXTURE_2D, meshes[i].textureID);
-	//
-	//	glActiveTexture(GL_TEXTURE0+1);
-	//	glBindTexture(GL_TEXTURE_2D, textures[2].textureID);
-	//
-	//	glBindVertexArray(meshes[i].vao);
-	//	glBindBuffer(GL_ARRAY_BUFFER, meshes[i].vbo);
-	//	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, meshes[i].ibo);
-	//
-	//	glDrawElements(GL_TRIANGLES, meshes[i].faceCount * 3, GL_UNSIGNED_SHORT, 0);
-	//}
-	
 
 }
 
@@ -172,7 +153,7 @@ std::vector<std::vector<float>> ContentManager::getMeshBoxes()
 {
 	std::vector<std::vector<float>> size;
 
-	for (int i = 0; i < meshes.size(); i++)
+	for (size_t i = 0; i < meshes.size(); i++)
 	{
 		size.push_back(meshes[i].getBoundaries());
 	}
